Fix critical TUI main loop bracket syntax error
- Fixed missing closing brace for match statement in pool loader message handling
- Fixed extra closing brace that was breaking the transaction message processing loop
- These syntax errors were preventing the TUI main loop from processing transaction updates from background tasks
- Background transaction updater and pool loader communication channels should now work properly

+++ src/pool_db.rs
@@ -144,12 +144,12 @@ impl PoolDatabase {
     /// Count pools in database
     pub fn pool_count(&self) -> Result<u32> {
         let conn = self.conn.lock().unwrap();
-        let count: u32 = conn.query_row(
+        let count: i64 = conn.query_row(
             "SELECT COUNT(*) FROM pools",
             [],
             |row| row.get(0),
         )?;
-        Ok(count)
+        Ok(count as u32)
     }
 
     /// Clear all pools (useful for updates)
